# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ROIFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpportunityInput {
  /** Short name for the task */
  title: string;
  /** Detailed description of the manual process */
  description: string;
  /** How many times per month this task is done */
  frequencyPerMonth: number;
  /** How many minutes each run takes */
  minutesPerRun: number;
  /** Number of people involved in the task */
  peopleInvolved: number;
}

export type OpportunityPriority =
  (typeof OpportunityPriority)[keyof typeof OpportunityPriority];

export const OpportunityPriority = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export interface Opportunity {
  id: number;
  title: string;
  description: string;
  frequencyPerMonth: number;
  minutesPerRun: number;
  peopleInvolved: number;
  taskSummary: string;
  trigger: string;
  appsInvolved: string[];
  outputAction: string;
  monthlyHoursLost: number;
  estimatedHoursSaved: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  roiScore: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  complexityScore: number;
  priority: OpportunityPriority;
  automationRecommendation: string;
  suggestedToolStack: string[];
  n8nWorkflowJson: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type ListOpportunitiesParams = {
  sort?: ListOpportunitiesSort;
  order?: ListOpportunitiesOrder;
};

export type ListOpportunitiesSort =
  (typeof ListOpportunitiesSort)[keyof typeof ListOpportunitiesSort];

export const ListOpportunitiesSort = {
  roi_score: "roi_score",
  priority: "priority",
  created_at: "created_at",
} as const;

export type ListOpportunitiesOrder =
  (typeof ListOpportunitiesOrder)[keyof typeof ListOpportunitiesOrder];

export const ListOpportunitiesOrder = {
  asc: "asc",
  desc: "desc",
} as const;