# lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ROIFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityPriority } from "./opportunityPriority";

export interface Opportunity {
  id: number;
  title: string;
  description: string;
  frequencyPerMonth: number;
  minutesPerRun: number;
  peopleInvolved: number;
  taskSummary: string;
  trigger: string;
  appsInvolved: string[];
  outputAction: string;
  monthlyHoursLost: number;
  estimatedHoursSaved: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  roiScore: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  complexityScore: number;
  priority: OpportunityPriority;
  automationRecommendation: string;
  suggestedToolStack: string[];
  n8nWorkflowJson: string;
  createdAt: Date;
}