# lib/api-zod/src/generated/types/opportunityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ROIFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpportunityInput {
  /** Short name for the task */
  title: string;
  /** Detailed description of the manual process */
  description: string;
  /** How many times per month this task is done */
  frequencyPerMonth: number;
  /** How many minutes each run takes */
  minutesPerRun: number;
  /** Number of people involved in the task */
  peopleInvolved: number;
}